backend/app/middleware/rate_limiting.py
"""Rate Limiting Middleware"""
from starlette.middleware.base import BaseHTTPMiddleware
from fastapi import Request
import logging

logger = logging.getLogger(__name__)

class RateLimitMiddleware(BaseHTTPMiddleware):
    """Middleware to enforce rate limiting"""
    
    async def dispatch(self, request: Request, call_next):
        # TODO: Implement rate limiting logic
        response = await call_next(request)
        return response
