backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# PostgreSQL Extensions
pgvector==0.2.4

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
openai==1.10.0
google-generativeai==0.3.2
tiktoken==0.5.2

# Vector Search
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# WhatsApp API
requests==2.31.0


# Storage
boto3==1.34.34
supabase==2.3.0

# Firebase (for notifications)
firebase-admin==6.4.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
phonenumbers==8.13.27

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

faker==22.6.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6

# Core FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Configuration Management
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests (for WhatsApp API)
requests==2.31.0


# Utilities
python-dateutil==2.8.2

# Testing (optional but recommended)
pytest==7.4.4
pytest-asyncio==0.23.3

# Core FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL (using psycopg instead of psycopg2-binary for Python 3.14 compatibility)
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.3.3

# Redis for caching
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Configuration Management
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests (for WhatsApp API)
requests==2.31.0


# Utilities
python-dateutil==2.8.2

# Testing (optional but recommended)
pytest==7.4.4
pytest-asyncio==0.23.3


